+++ matrix_client/api_async.py
@@ -4,6 +4,6 @@ if sys.version_info < (3, 5):
     raise ValueError("The asyncio version of the api "
                      "is only supported on Python 3.5+")  # pragma: nocover
 else:
-    from ._api_asyncio import AsyncHTTPAPI
+    from ._api_async import AsyncHTTPAPI
 
     __all__ = ['AsyncHTTPAPI']
